change answer 0

+++ 0x02-functions_nested_loops/0-holberton.c
@@ -1,10 +1,16 @@
 #include "holberton.h"
 
+/**
+ * main - Entry point
+ *
+ * Return: Always 0 (Success)
+ */
 int main(void)
 {
 	char *c = "Holberton";
-	int i;
-	for(i = 0; i <= sizeof(c); i++)
+	unsigned long i;
+
+	for (i = 0; i <= sizeof(c); i++)
 		_putchar(c[i]);
 	_putchar('\n');
 	return (0);
